Incorporated ShoppingService and finalized domain (/w ID)

+++ PlatformServices/src/main/java/com/infosupport/t2c3/data/BasicRepository.java
@@ -10,6 +10,6 @@ import org.springframework.stereotype.Repository;
  * @param <T> The class of the Model
  */
 @Repository
-public interface BasicRepository<T> extends JpaRepository<T, Integer> {
+public interface BasicRepository<T> extends JpaRepository<T, Long> {
 
 }
